Divide workspace to the two files
